Added input_test.py and snake example
input_test.py tests pyg-engine's input system, demonstrating reading the
inputs, and creating input events. Currently the pyg-engine backings
aren't developed for this, but this example will be used to develop the
backings. Really, what I'm doing is doing is writing the way I want to
make programs on pyg-engine, and I'm going to develop the backings to
support it. Kind of backwards, but it serves as an end goal to what I
want this to eventually be. The snake example is for the same purpose
too.